Clippy Fixes

+++ src/styles/rich.rs
@@ -16,10 +16,10 @@ pub(crate) fn pulse(ncols: i16, current_time: f32) -> String {
         "#f42670", "#f92672", "#f42670", "#e6276c", "#d12a66", "#b72c5e", "#993056", "#7b334d",
         "#613545", "#4c383f",
     ]
-    .repeat((ncols as f32 / 18 as f32) as usize + 2);
+    .repeat((ncols as f32 / 18_f32) as usize + 2);
 
     let pulse_len = pulse.len();
-    let offset = (-current_time * 15 as f32) as i16 % 18;
+    let offset = (-current_time * 15_f32) as i16 % 18;
     let mut pulse_string = String::new();
 
     for i in offset..(offset + ncols) {
